Cleaned up code formatting

+++ main/src/georegression/metric/Distance2D_F64.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021, Peter Abeles. All Rights Reserved.
+ * Copyright (C) 2022, Peter Abeles. All Rights Reserved.
  *
  * This file is part of Geometric Regression Library (GeoRegression).
  *
@@ -30,7 +30,6 @@ import georegression.struct.shapes.Quadrilateral_F64;
 import georegression.struct.shapes.Rectangle2D_F64;
 import org.jetbrains.annotations.Nullable;
 
-
 /**
  * Functions related to finding the distance of one shape from another shape. This is often
  * closely related to finding the {@link ClosestPoint3D_F64 closest point}.
@@ -40,7 +39,6 @@ import org.jetbrains.annotations.Nullable;
 // TODO distance between two line segments, line lines
 // handle parallel overlapping cases by returning zero
 public class Distance2D_F64 {
-
 
 	/**
 	 * <p>
@@ -65,8 +63,8 @@ public class Distance2D_F64 {
 	 * @param y Point's y-coordinate
 	 * @return Distance the closest point on the line is away from the point.
 	 */
-	public static double distance( LineParametric2D_F64 line, double x , double y ) {
-		return Math.sqrt(distanceSq(line, x,y));
+	public static double distance( LineParametric2D_F64 line, double x, double y ) {
+		return Math.sqrt(distanceSq(line, x, y));
 	}
 
 	/**
@@ -79,17 +77,17 @@ public class Distance2D_F64 {
 	 * @return Euclidean distance squared to the closest point on the line is away from the point.
 	 */
 	public static double distanceSq( LineParametric2D_F64 line, Point2D_F64 p ) {
-		double scale = Math.max(Math.abs(line.slope.x),Math.abs(line.slope.y));
+		double scale = Math.max(Math.abs(line.slope.x), Math.abs(line.slope.y));
 
-		double t = ClosestPoint2D_F64.closestPointT( line, p , scale);
+		double t = ClosestPoint2D_F64.closestPointT(line, p, scale);
 
-		double a = (line.slope.x/scale) * t + line.p.x;
-		double b = (line.slope.y/scale) * t + line.p.y;
+		double a = (line.slope.x/scale)*t + line.p.x;
+		double b = (line.slope.y/scale)*t + line.p.y;
 
 		double dx = p.x - a;
 		double dy = p.y - b;
 
-		return dx * dx + dy * dy;
+		return dx*dx + dy*dy;
 	}
 
 	/**
@@ -102,18 +100,18 @@ public class Distance2D_F64 {
 	 * @param y Point's y-coordinate
 	 * @return Euclidean distance squared to the closest point on the line is away from the point.
 	 */
-	public static double distanceSq( LineParametric2D_F64 line, double x , double y ) {
-		double scale = Math.max(Math.abs(line.slope.x),Math.abs(line.slope.y));
+	public static double distanceSq( LineParametric2D_F64 line, double x, double y ) {
+		double scale = Math.max(Math.abs(line.slope.x), Math.abs(line.slope.y));
 
-		double t = ClosestPoint2D_F64.closestPointT( line, x, y , scale);
+		double t = ClosestPoint2D_F64.closestPointT(line, x, y, scale);
 
-		double a = (line.slope.x/scale) * t + line.p.x;
-		double b = (line.slope.y/scale) * t + line.p.y;
+		double a = (line.slope.x/scale)*t + line.p.x;
+		double b = (line.slope.y/scale)*t + line.p.y;
 
 		double dx = x - a;
 		double dy = y - b;
 
-		return dx * dx + dy * dy;
+		return dx*dx + dy*dy;
 	}
 
 	/**
@@ -139,8 +137,8 @@ public class Distance2D_F64 {
 	 * @param y Point's y-coordinate
 	 * @return Euclidean distance of the closest point on a line is away from a point.
 	 */
-	public static double distance( LineSegment2D_F64 line, double x , double y) {
-		return Math.sqrt(distanceSq(line, x,y));
+	public static double distance( LineSegment2D_F64 line, double x, double y ) {
+		return Math.sqrt(distanceSq(line, x, y));
 	}
 
 	/**
@@ -156,18 +154,18 @@ public class Distance2D_F64 {
 		double a = line.b.x - line.a.x;
 		double b = line.b.y - line.a.y;
 
-		double t = a * ( p.x - line.a.x ) + b * ( p.y - line.a.y );
-		t /= ( a * a + b * b );
+		double t = a*(p.x - line.a.x) + b*(p.y - line.a.y);
+		t /= (a*a + b*b);
 
 		// if the point of intersection is past the end points return the distance
 		// from the closest end point
-		if( t < 0 ) {
+		if (t < 0) {
 			return UtilPoint2D_F64.distanceSq(line.a.x, line.a.y, p.x, p.y);
-		} else if( t > 1.0 )
+		} else if (t > 1.0)
 			return UtilPoint2D_F64.distanceSq(line.b.x, line.b.y, p.x, p.y);
 
 		// return the distance of the closest point on the line
-		return UtilPoint2D_F64.distanceSq(line.a.x + t * a, line.a.y + t * b, p.x, p.y);
+		return UtilPoint2D_F64.distanceSq(line.a.x + t*a, line.a.y + t*b, p.x, p.y);
 	}
 
 	/**
@@ -180,42 +178,43 @@ public class Distance2D_F64 {
 	 * @param y Point's y-coordinate
 	 * @return Euclidean distance squared of the closest point on a line is away from a point.
 	 */
-	public static double distanceSq( LineSegment2D_F64 line, double x , double y ) {
+	public static double distanceSq( LineSegment2D_F64 line, double x, double y ) {
 		double a = line.b.x - line.a.x;
 		double b = line.b.y - line.a.y;
 
-		double t = a * ( x - line.a.x ) + b * ( y - line.a.y );
-		t /= ( a * a + b * b );
+		double t = a*(x - line.a.x) + b*(y - line.a.y);
+		t /= (a*a + b*b);
 
 		// if the point of intersection is past the end points return the distance
 		// from the closest end point
-		if( t < 0 ) {
+		if (t < 0) {
 			return UtilPoint2D_F64.distanceSq(line.a.x, line.a.y, x, y);
-		} else if( t > 1.0 )
+		} else if (t > 1.0)
 			return UtilPoint2D_F64.distanceSq(line.b.x, line.b.y, x, y);
 
 		// return the distance of the closest point on the line
-		return UtilPoint2D_F64.distanceSq(line.a.x + t * a, line.a.y + t * b, x, y);
+		return UtilPoint2D_F64.distanceSq(line.a.x + t*a, line.a.y + t*b, x, y);
 	}
-
 
 	/**
 	 * Finds the distance between the two line segments
+	 *
 	 * @param segmentA Line segment. Not modified.
 	 * @param segmentB Line segment. Not modified.
 	 * @return Euclidean distance of the closest point between the two line segments.
 	 */
-	public static double distance( LineSegment2D_F64 segmentA , LineSegment2D_F64 segmentB ) {
+	public static double distance( LineSegment2D_F64 segmentA, LineSegment2D_F64 segmentB ) {
 		return Math.sqrt(distanceSq(segmentA, segmentB));
 	}
 
 	/**
 	 * Finds the distance squared between the two line segments
+	 *
 	 * @param segmentA Line segment. Not modified.
 	 * @param segmentB Line segment. Not modified.
 	 * @return Euclidean distance squared of the closest point between the two line segments.
 	 */
-	public static double distanceSq( LineSegment2D_F64 segmentA , LineSegment2D_F64 segmentB ) {
+	public static double distanceSq( LineSegment2D_F64 segmentA, LineSegment2D_F64 segmentB ) {
 
 		// intersection of the two lines relative to A
 		double slopeAX = segmentA.slopeX();
@@ -223,27 +222,27 @@ public class Distance2D_F64 {
 		double slopeBX = segmentB.slopeX();
 		double slopeBY = segmentB.slopeY();
 
-		double ta = slopeBX*( segmentA.a.y - segmentB.a.y ) - slopeBY*( segmentA.a.x - segmentB.a.x );
+		double ta = slopeBX*(segmentA.a.y - segmentB.a.y) - slopeBY*(segmentA.a.x - segmentB.a.x);
 		double bottom = slopeBY*slopeAX - slopeAY*slopeBX;
 
 		// see they intersect
-		if( bottom != 0 ) {
+		if (bottom != 0) {
 			// see if the intersection is inside of lineA
 			ta /= bottom;
-			if( ta >= 0 && ta <= 1.0 ) {
+			if (ta >= 0 && ta <= 1.0) {
 				// see if the intersection is inside of lineB
-				double tb = slopeAX*( segmentB.a.y - segmentA.a.y ) - slopeAY*( segmentB.a.x - segmentA.a.x );
+				double tb = slopeAX*(segmentB.a.y - segmentA.a.y) - slopeAY*(segmentB.a.x - segmentA.a.x);
 				tb /= slopeAY*slopeBX - slopeBY*slopeAX;
-				if( tb >= 0 && tb <= 1.0 )
+				if (tb >= 0 && tb <= 1.0)
 					return 0;
 			}
 		}
 
 		double closest = Double.MAX_VALUE;
-		closest = Math.min(closest,distanceSq(segmentA, segmentB.a));
-		closest = Math.min(closest,distanceSq(segmentA, segmentB.b));
-		closest = Math.min(closest,distanceSq(segmentB, segmentA.a));
-		closest = Math.min(closest,distanceSq(segmentB, segmentA.b));
+		closest = Math.min(closest, distanceSq(segmentA, segmentB.a));
+		closest = Math.min(closest, distanceSq(segmentA, segmentB.b));
+		closest = Math.min(closest, distanceSq(segmentB, segmentA.a));
+		closest = Math.min(closest, distanceSq(segmentB, segmentA.b));
 
 		return closest;
 	}
@@ -255,8 +254,8 @@ public class Distance2D_F64 {
 	 * @param p Point
 	 * @return Distance apart
 	 */
-	public static double distance( Quadrilateral_F64 quad , Point2D_F64 p ) {
-		return Math.sqrt(distanceSq(quad,p));
+	public static double distance( Quadrilateral_F64 quad, Point2D_F64 p ) {
+		return Math.sqrt(distanceSq(quad, p));
 	}
 
 	/**
@@ -266,7 +265,7 @@ public class Distance2D_F64 {
 	 * @param p Point
 	 * @return Distance squared apart
 	 */
-	public static double distanceSq( Quadrilateral_F64 quad , Point2D_F64 p ) {
+	public static double distanceSq( Quadrilateral_F64 quad, Point2D_F64 p ) {
 		LineSegment2D_F64 seg = LineSegment2D_F64.wrap(quad.a, quad.b);
 		double a = distanceSq(seg, p);
 		seg.a = quad.b;seg.b = quad.c;
@@ -284,7 +283,7 @@ public class Distance2D_F64 {
 	 * @param p Point
 	 * @return Distance squared apart
 	 */
-	public static double distance( Polygon2D_F64 poly , Point2D_F64 p ) {
+	public static double distance( Polygon2D_F64 poly, Point2D_F64 p ) {
 		return Math.sqrt(distanceSq(poly, p, null));
 	}
 
@@ -296,19 +295,19 @@ public class Distance2D_F64 {
 	 * @param storage Optional storage for linesegment which is used internally to compute the distance
 	 * @return Distance squared apart
 	 */
-	public static double distanceSq( Polygon2D_F64 poly , Point2D_F64 p , @Nullable LineSegment2D_F64 storage ) {
-		if( storage == null )
+	public static double distanceSq( Polygon2D_F64 poly, Point2D_F64 p, @Nullable LineSegment2D_F64 storage ) {
+		if (storage == null)
 			storage = new LineSegment2D_F64();
 
 		double minimum = Double.MAX_VALUE;
 		for (int i = 0; i < poly.size(); i++) {
-			int j = (i+1)%poly.size();
+			int j = (i + 1)%poly.size();
 
 			storage.a.setTo(poly.vertexes.data[i]);
 			storage.b.setTo(poly.vertexes.data[j]);
 
 			double d = distanceSq(storage, p);
-			if( d < minimum )
+			if (d < minimum)
 				minimum = d;
 		}
 
@@ -324,8 +323,8 @@ public class Distance2D_F64 {
 	 * @param p The point. Not modified.
 	 * @return Euclidean distance of the closest point on the line to the specified point.
 	 */
-	public static double distance( LineGeneral2D_F64 line , Point2D_F64 p ) {
-		return Math.abs(line.A*p.x + line.B*p.y + line.C) / Math.sqrt( line.A*line.A + line.B*line.B );
+	public static double distance( LineGeneral2D_F64 line, Point2D_F64 p ) {
+		return Math.abs(line.A*p.x + line.B*p.y + line.C)/(double)Math.sqrt(line.A*line.A + line.B*line.B);
 	}
 
 	/**
@@ -338,12 +337,13 @@ public class Distance2D_F64 {
 	 * @param p The point. Not modified.
 	 * @return Euclidean distance of the closest point on the line to the specified point.
 	 */
-	public static double distanceNorm(LineGeneral2D_F64 line, Point2D_F64 p) {
+	public static double distanceNorm( LineGeneral2D_F64 line, Point2D_F64 p ) {
 		return Math.abs(line.A*p.x + line.B*p.y + line.C);
 	}
 
 	/**
 	 * Returns the distance of the closest point on the line from the origin
+	 *
 	 * @param line Line
 	 * @return Euclidean distance
 	 */
@@ -360,7 +360,7 @@ public class Distance2D_F64 {
 	 * @param p Point
 	 * @return Euclidean distance
 	 */
-	public static double distance(EllipseRotated_F64 ellipse , Point2D_F64 p ) {
+	public static double distance( EllipseRotated_F64 ellipse, Point2D_F64 p ) {
 		return Math.sqrt(distance2(ellipse, p));
 	}
 
@@ -371,7 +371,7 @@ public class Distance2D_F64 {
 	 * @param p Point
 	 * @return Euclidean distance squared
 	 */
-	public static double distance2(EllipseRotated_F64 ellipse , Point2D_F64 p ) {
+	public static double distance2( EllipseRotated_F64 ellipse, Point2D_F64 p ) {
 		// put point into ellipse's reference frame
 		double cphi = Math.cos(ellipse.phi);
 		double sphi = Math.sin(ellipse.phi);
@@ -380,7 +380,7 @@ public class Distance2D_F64 {
 		double yc = p.y - ellipse.center.y;
 		double r = Math.sqrt(xc*xc + yc*yc);
 
-		double x =  cphi*xc + sphi*yc;
+		double x = cphi*xc + sphi*yc;
 		double y = -sphi*xc + cphi*yc;
 
 		double ct = x/r;
@@ -389,7 +389,7 @@ public class Distance2D_F64 {
 		x = ellipse.center.x + ellipse.a*ct*cphi - ellipse.b*st*sphi;
 		y = ellipse.center.y + ellipse.a*ct*sphi + ellipse.b*st*cphi;
 
-		return p.distance2(x,y);
+		return p.distance2(x, y);
 	}
 
 	/**
@@ -399,8 +399,8 @@ public class Distance2D_F64 {
 	 * @param b (Input) rectangle
 	 * @return IoU score
 	 */
-	public static double scoreIoU(Rectangle2D_F64 a, Rectangle2D_F64 b ) {
-		double areaI = Intersection2D_F64.intersectionArea(a,b);
+	public static double scoreIoU( Rectangle2D_F64 a, Rectangle2D_F64 b ) {
+		double areaI = Intersection2D_F64.intersectionArea(a, b);
 
 		if (areaI == 0.0)
 			return 0.0;
@@ -415,8 +415,8 @@ public class Distance2D_F64 {
 	 * @param b (Input) Simple polygon
 	 * @return IoU score
 	 */
-	public static double scoreIoU(Polygon2D_F64 a, Polygon2D_F64 b,
-								  @Nullable AreaIntersectionPolygon2D_F64 computeArea) {
+	public static double scoreIoU( Polygon2D_F64 a, Polygon2D_F64 b,
+								   @Nullable AreaIntersectionPolygon2D_F64 computeArea ) {
 		if (computeArea == null)
 			computeArea = new AreaIntersectionPolygon2D_F64();
